+++ src/map/in_neighborhood/mod.rs
@@ -1,20 +1,13 @@
-use std::fmt::Debug;
-
-use in_byte::backend::NeighborhoodArray;
-
+pub mod in_array;
 pub mod in_byte;
 pub mod in_vec;
 
-pub trait InNeighborhood<'a, 'b>:
-    From<NeighborhoodArray> + TryInto<NeighborhoodArray>
-    + Debug + Default
-where
-    'b: 'a,
+pub trait InNeighborhood<'a, 'b: 'a>:
 {
-    type Iter: Iterator<Item = &'b (usize, usize)> + 'a;
-    fn min(&'a self) -> Option<&'a usize>;
-    fn max(&'a self) -> Option<&'a usize>;
-    fn push(&mut self, value: usize);
+    type Iter: Iterator<Item = &'b (usize, usize)>;
+    fn min(&'a self) -> Option<&'b usize>;
+    fn max(&'a self) -> Option<&'b usize>;
+    fn push(&'a mut self, value: usize);
     fn iter(&'a self) -> Self::Iter;
     fn gca_skip(left: &'a Self, right: &'a Self) -> bool {
         match (left.max(), right.min()) {
